+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,35 @@
 # next-docs-ui
 
+## 12.1.0
+
+### Minor Changes
+
+- 0a377a9: **Pass the `icon` prop to code blocks as HTML instead of MDX attribute.**
+
+  **why:** Only MDX flow elements support attributes with JSX value, like:
+
+  ```mdx
+  <Pre icon={<svg />}>...</Pre>
+  ```
+
+  As Shiki outputs hast elements, we have to convert the output of Shiki to a MDX flow element so that we can pass the `icon` property.
+
+  Now, `rehype-code` passes a HTML string instead of JSX, and render it with `dangerouslySetInnerHTML`:
+
+  ```mdx
+  <Pre icon="<svg />">...</Pre>
+  ```
+
+  **migrate:** Not needed, it should work seamlessly.
+
+### Patch Changes
+
+- 0a377a9: Close sidebar on collapse
+- 5f86faa: Improve multi-line code blocks
+- Updated dependencies [0a377a9]
+- Updated dependencies [0a377a9]
+  - fumadocs-core@12.1.0
+
 ## 12.0.7
 
 ### Patch Changes